+++ common/src/main/java/pepjebs/mapatlases/utils/TriState.java
@@ -1,0 +1,7 @@
+package pepjebs.mapatlases.utils;
+
+public enum TriState {
+    PASS,
+    SET_TRUE,
+    SET_FALSE
+}